+++ src/main/java/com/example/demo/model/Category.java
@@ -12,7 +12,7 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(unique = true, nullable = false)
     private String name;
@@ -22,7 +22,7 @@ public class Category {
 
     public Category() {}
 
-    public Category(Long id, String name, String type) {
+    public Category(int id, String name, String type) {
         this.id = id;
         this.name = name;
         this.type = type;
@@ -34,11 +34,11 @@ public class Category {
         }
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
